refactor: move riscv features code to `kstd`
